requirements.txt
python-dotenv>=1.0.0
requests>=2.28.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.100.0
psycopg2-binary>=2.9.0
flask>=3.0.0
flask-cors>=4.0.0
